requirements.txt
paho-mqtt==2.1.0
ujson==5.11.0
